Fix: scroll
Creé custom hook para centralizar lógica de scrollintoview
